Temporarily ignore stream lengths by indirect object reference.
PDF allows streams to specify the length of the string using an indirect
object reference like "20 0 R" instead of directly storing the length in
the stream dictionary.  This is an important capability, allowing a PDF
file to be generated on the fly, outputting a forward references to the
stream length followed by the contents of the stream itself (before the
length of that stream is actually known), followed by the referenced
indirect object containing the actual length of the stream.

Unfortunately, this type of string length complicates the PDF::Data
parser, which is trying to parse all the PDF objects in a single pass
instead of relying on the cross-reference table of byte offsets, in case
those byte offsets are wrong.  Because the indirect object references
are resolved as a later step after the stream parsing, the actual stream
length stored in the indirect object is not readily available to this
parsing code.

Although a reference to a previously-parsed indirect object is available
to this parsing code, it's nearly certain that any stream length stored
as an indirect object reference will be a forward reference to a later
object that hasn't been parsed yet.  (After all, there's no reason to
use an indirect object reference for the stream length if that length
was already known before starting to output the stream data!)

This could be solved by coming up with another way to access the value
of that indirect object reference before parsing is complete, either by
relying on the byte offsets in the cross-reference table (assuming the
offsets are correct), or perhaps by using a regex match across all of
the PDF data to find the indirect object if necessary.

At this point in the code, the $length variable will contain a reference
to a scalar value containing the object number for the indirect object
reference, such as \20 for a reference like "20 0 R".  Until a solution
is implemented as described above to retrieve this value on the fly
during this parsing step, just silently ignore the specified indirect
object reference for now, and pretend the stream length is missing.

Without knowing the length of the stream data to expect, this triggers
the code to search for the "endstream" marker to automatically determine
the stream length.  This isn't perfect; without knowing the exact stream
length to expect, it's impossible to know for certain whether or not the
stream data ends with a newline, and hypothetically the next "endstream"
marker could actually be part of the valid stream data.  However, this
would only be likely to happen if the stream data actually contains
another full independent PDF file as uncompressed stream data inside the
PDF file for some reason.

This parsing code should be fixed to use the correct stream length in
the referenced indirect object, but automatic detection of the stream
length seems to work well in practice.