Replace extremely inefficient regular expression with rindex().
The regular expression /(.*)\z/ is a simple way to get the last partial
line of the buffer, but in practice it is extremely inefficient.  Using
Devel::NYTProf to profile a particular test case, it was spending over
8 MINUTES of CPU time just matching this regular expression repeatedly,
for a total of 53,587 times.

Since the only purpose of this regex match was to calculate the length
of the final partial line in the buffer to wrap to a new line before
exceeding 255 characters in length, this commit fixes the inefficiency
by replacing the regular expression with a call to rindex() to find the
last newline in the string, and using that offset for the calculation.

Profiling after this change with the same test case, running this line
of code 53,387 times now takes a total of 41.3 milliseconds instead of
486 seconds.  That saved 99.99% of the CPU time, making this line of
code 11,767 times faster than the original, for this test case.

As an experiment, I also tried profiling a modified version of the
original regular expression, adding an independent group to reduce the
amount of backtracking required: /((?>.*))\z/

I thought this would still be very inefficient, but it seemed like it
should reduce backtracking and thus optimize the regular expression to
some degree.  However, when I profiled this change on the same test
case, the CPU time spent matching the regular expression inexplicably
INCREASED from 486 seconds (8 minutes) to 2,820 seconds (48 minutes)
somehow!

Using Regexp::Debugger, I confirmed that the new regular expression did
indeed reduce backtracking, requiring 249 steps in the regex engine for
the new regex instead of 412 steps for the original.  Apparently some of
the new steps are much more expensive than the ones being skipped,
causing the CPU time to increase despite reducing the number of steps
required for the regex to actually match.

I also profiled a different regular expression which anchors the start
of the match as well as the end: /\A(?>(?:.*\n)*)(.*)\z/

This version actually does reduce the total CPU time, from 486 seconds
for the original to 200 seconds for this regular expression.  The first
match requires only 40 steps in the regex engine, rather than 412 steps
for the original.  This version is over twice as fast as the original,
but still very slow.

I next tried the following regular expression, because I had a hunch
that it would be much faster: /\A(?s:.*\n)?(.*)\z/

This version is blindingly fast compared to all of the other regular
expressions, bringing the total CPU time for the 53,587 matches for this
test case down to 226 milliseconds!  Still not as efficient as rindex()
but finally in the ballpark!  The number of steps for this regular
expression is down to 17 steps now.  I also realized that the anchors
aren't actually required here, because the anchors are always guaranteed
to be satisfied on the first match.  If I remove the \A and \z anchors,
the first match drops from 17 steps to 13 steps, and the total CPU time
drops slightly from 226 milliseconds to 222 milliseconds.

Regardless, the rindex() approach is still the best solution here, at
only 41.3 milliseconds for the same test case, so I'm going with that!