Accept and parse valid PDF data regardless of leading/trailing garbage.
On second thought, there's no need to apply the same limits as Adobe
Acrobat does.  If the file contains valid PDF data anywhere in the file,
with an unlimited amount of potential leading and/or trailing garbage,
just parse the PDF data and ignore the rest of the file.  If the parsed
data is saved again, a proper valid PDF will be constructed from the
data parsed out of the original malformed PDF file.